+++ pkg/codegen/nodejs/gen.go
@@ -1221,7 +1221,8 @@ type npmPackage struct {
 }
 
 type npmPulumiManifest struct {
-	Resource bool `json:"resource,omitempty"`
+	Resource    bool   `json:"resource,omitempty"`
+	ProviderURL string `json:"providerURL,omitempty"`
 }
 
 func genNPMPackageMetadata(pkg *schema.Package, info NodePackageInfo) string {
@@ -1253,7 +1254,8 @@ func genNPMPackageMetadata(pkg *schema.Package, info NodePackageInfo) string {
 		},
 		DevDependencies: devDependencies,
 		Pulumi: npmPulumiManifest{
-			Resource: true,
+			Resource:    true,
+			ProviderURL: pkg.ProviderURL,
 		},
 	}