+++ CHANGELOG.md
@@ -3,8 +3,11 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 
+- Add providerURL field to package definition
+  [#4947](https://github.com/pulumi/pulumi/pull/4947)
+
 - Go program gen: Improved handling for pulumi.Map types
-  [#491](https://github.com/pulumi/pulumi/pull/4914)
+  [#4914](https://github.com/pulumi/pulumi/pull/4914)
 
 - Go SDK: Input type interfaces should declare pointer type impls where appropriate
   [#4911](https://github.com/pulumi/pulumi/pull/4911)